grovia-backend/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
mysqlclient==2.2.0
pymysql==1.1.0
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pycryptodome==3.20.0

# Machine Learning & Computer Vision
numpy==1.26.4
pillow==10.2.0
opencv-python==4.9.0.80
scikit-image==0.22.0

# Google Gemini AI - FREE & Powerful!
google-generativeai==0.3.2

# Cloud Storage - Cloudinary for image uploads
cloudinary==1.36.0

# Data Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Testing (optional)
pytest==7.4.0
pytest-asyncio==0.23.4
httpx==0.26.0
